html/semantics/embedded-content-0/media-elements/loading-the-media-resource/026.html

<!doctype html>
<title>invoking resource selection with implicit pause() by removing from document</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
var v;
var t = async_test(function(t) {
  v = document.createElement('video');
  document.body.appendChild(v);
  v.setAttribute('src', ''); // invokes resource selection
  assert_equals(v.networkState, v.NETWORK_NO_SOURCE, 'networkState before removing src');
  v.removeAttribute('src'); // no side-effects
  assert_equals(v.networkState, v.NETWORK_NO_SOURCE, 'networkState after removing src');
  v.parentNode.removeChild(v); // implicit pause() which ... does nothing
  assert_equals(v.networkState, v.NETWORK_NO_SOURCE, 'networkState after removing v');
});
</script>
<script>
t.step(function() {
  assert_equals(v.networkState, v.NETWORK_EMPTY, 'networkState in separate script');
  t.done();
});
</script>
